refactor: don't import pci from arch
